ATmega2_ISPProgrammer is the port of ArduinoISP sketch, it uses interrupt based serial communication functions (19200 bps), pwm functions and SPI functions of ATmegaCLib2 library, and can be compiled for any ATmega microcontroller supported by the library (and probably many more).
The status LEDs are user definable but the HearBeat LED must be on the PWM channel No1, as defined by Arduino library. This is configured automatically inside library depending on the target microcontroller, but you must know which pin is that (look inside library).

Currently, this was tested on ATmega644P, ATmega328P and ATmega168P (this one currently is not supported by Arduino IDE and toolchain provided in the original Arduino package to date). ATmega644P was part of "8Bit Offense" board which have all the required LEDs to work as a complete ISP Programmer (yes, the microcontroller is overkill but you can do it if you don't have anything else at hand).

With this, you (and me) no longer depend on Arduino IDE.